requirements.txt
altair==5.1.2
numpy==1.26.1
pandas==2.1.2
pydeck==0.8.0
streamlit==1.28.1
duckdb==1.1.3
python-dotenv==1.0.1
plotly==5.24.1